[Chore/#11] enum 파일 분리
