+++ Gongbaek_iOS/Gongbaek_iOS/Global/Component/Button/CounterButton.swift
@@ -7,11 +7,6 @@
 
 import SwiftUI
 
-enum CounterButtonState {
-    case increase
-    case decrease
-}
-
 struct CounterButton: View {
     var state: CounterButtonState
     var onTap: (() -> Void)?
