+++ Gongbaek_iOS/Gongbaek_iOS/Global/Component/Button/CycleButton.swift
@@ -7,34 +7,6 @@
 
 import SwiftUI
 
-enum CycleState {
-    case all
-    case weekly
-    case once
-    
-    var titleText: String {
-        switch self {
-        case .all: return "모든 주기"
-        case .weekly: return "매주 봐요"
-        case .once: return "한번만 봐요"
-        }
-    }
-    
-    var backgroundColor: Color {
-        switch self {
-        case .all: return .gray01
-        case .weekly, .once: return .gray09
-        }
-    }
-    
-    var foregroundColor: Color {
-        switch self {
-        case .all: return .gray10
-        case .weekly, .once: return .grayWhite
-        }
-    }
-}
-
 struct CycleButton: View {
     var state: CycleState
     var onTap: (() -> Void)?
