+++ Gongbaek_iOS/Gongbaek_iOS/Global/Enum/CounterButtonState.swift
@@ -1,0 +1,13 @@
+//
+//  CounterButtonState.swift
+//  Gongbaek_iOS
+//
+//  Created by 김민서 on 1/15/25.
+//
+
+import SwiftUI
+
+enum CounterButtonState {
+    case increase
+    case decrease
+}
